3d_to_image/requirements.txt
matplotlib==3.9.2
numpy==2.0.1
pandas==2.2.3
Pillow==11.1.0
pytorch3d==0.7.8
pyvista==0.44.2
PyYAML==6.0.2
scikit_learn==1.6.1
seaborn==0.13.2
torch==2.4.1
torchvision==0.19.1
tqdm==4.67.1